split system + material component
