+++ src/systems/index.js
@@ -2,3 +2,4 @@ require('./camera');
 require('./geometry');
 require('./material');
 require('./light');
+require('./texture');
